Projeto_IA/requirements.txt
requests
SpeechRecognition
gTTS
pygame
deep-translator
numpy
pyttsx3
Pillow
pyaudio